[pdpix] Don't Optimize Size of Structures

+++ examples/pipe-open/server.rs
@@ -80,10 +80,10 @@ impl PipeServer {
     fn handle_fail(&mut self, qr: &demi_qresult_t) -> Result<usize> {
         let qd: QDesc = qr.qr_qd.into();
         let qt: QToken = qr.qr_qt.into();
-        let errno: i32 = qr.qr_ret;
+        let errno: i64 = qr.qr_ret;
 
         // Check if client has reset the connection.
-        if errno == libc::ECONNRESET {
+        if errno == libc::ECONNRESET as i64 {
             println!("INFO: client reset connection (qd={:?})", qd);
         } else {
             println!(
