+++ src/rust/catpowder/interop.rs
@@ -77,7 +77,7 @@ pub fn pack_result(rt: Rc<LinuxRuntime>, result: OperationResult, qd: QDesc, qt:
                     qr_opcode: demi_opcode_t::DEMI_OPC_FAILED,
                     qr_qd: qd.into(),
                     qr_qt: qt,
-                    qr_ret: e.errno,
+                    qr_ret: e.errno as i64,
                     qr_value: unsafe { mem::zeroed() },
                 }
             },
@@ -95,7 +95,7 @@ pub fn pack_result(rt: Rc<LinuxRuntime>, result: OperationResult, qd: QDesc, qt:
                 qr_opcode: demi_opcode_t::DEMI_OPC_FAILED,
                 qr_qd: qd.into(),
                 qr_qt: qt,
-                qr_ret: e.errno,
+                qr_ret: e.errno as i64,
                 qr_value: unsafe { mem::zeroed() },
             }
         },
